# omcgo/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OMC API
  description: Operations Management and Control for Base Stations
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/provisioning/tasks:
    get:
      tags: [Provisioning]
      summary: List provisioning tasks
      operationId: listProvisioningTasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated provisioning task list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProvisioningTask"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Provisioning]
      summary: Create a provisioning task
      operationId: createProvisioningTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProvisioningTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningTask"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/provisioning/tasks/{id}:
    get:
      tags: [Provisioning]
      summary: Get provisioning task by ID
      operationId: getProvisioningTask
      parameters:
        - $ref: "#/components/parameters/UUIDPathParam"
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningTask"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/provisioning/tasks/{id}/retry:
    post:
      tags: [Provisioning]
      summary: Retry a provisioning task
      operationId: retryProvisioningTask
      parameters:
        - $ref: "#/components/parameters/UUIDPathParam"
      responses:
        "200":
          description: Task retried
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningTask"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===========================================================================
  # Groups
  # ===========================================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
    UUIDPathParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request parameters"
        details:
          type: string
          description: Additional error details
        request_id:
          type: string
          description: Request tracking ID
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          description: Total number of records
          example: 150
        page:
          type: integer
          description: Current page number
          example: 1
        page_size:
          type: integer
          description: Number of items per page
          example: 20

    # ---- Auth ----
...
    ProvisioningTask:
      type: object
      properties:
        id:
          type: string
          format: uuid
        device_id:
          type: string
          format: uuid
        task_type:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        parameters:
          type: object
          additionalProperties: true
        result:
          type: object
          additionalProperties: true
        error_message:
          type: string
        retry_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    CreateProvisioningTaskRequest:
      type: object
      required: [device_id, task_type]
      properties:
        device_id:
          type: string
          format: uuid
        task_type:
          type: string
        parameters:
          type: object
          additionalProperties: true

    # ---- Groups ----
...